Extract service classes for importing role composites
